Renamed folder for fetching API URL
