+++ src/hooks/venues/searchHook.jsx
@@ -1,5 +1,5 @@
 import { Search } from '../../components/Search';
-import { getVenue } from '../../services/api/apiBase';
+import { getVenue } from '../../services/authorization/apiBase';
 import { useApiCall } from '../api/useApiCall';
 
 export const VenueSearch = () => {
