+++ src/pages/VenuePage/DetailsPage/index.jsx
@@ -1,6 +1,6 @@
 import { useParams } from 'react-router-dom';
 import { useApiCall } from '../../../hooks/api/useApiCall';
-import { getVenue } from '../../../services/Api/apiBase';
+import { getVenue } from '../../../services/authorization/apiBase';
 import { VenueDetails } from '../../../components/venue/venueData';
 
 export function DetailsPage() {
